+++ lang/cs/layouts.php
@@ -3,14 +3,13 @@
 return [
 
     'email.anonymous' => '-',
+    'modal.description' => 'Vaše zpětná vazba nám pomáhá se zlepšovat. Podělte se o své dojmy.',
     'email.from' => 'Od',
+    'modal.link' => 'Poslat zpětnou vazbu',
     'email.no_message' => 'Bez zprávy',
     'email.no_subject' => 'Bez předmětu',
     'email.sent_from' => 'Odesláno z',
-    'email.subject' => 'Zpětná vazba :app_www',
-    'modal.description' => 'Vaše zpětná vazba nám pomáhá se zlepšovat. Podělte se o své dojmy.',
-    'modal.link_text' => 'Zpětná vazba',
-    'modal.success_message' => 'Děkujeme za vaši zpětnou vazbu!',
-    'modal.title' => 'Poslat zpětnou vazbu',
+    'modal.success' => 'Děkujeme, vaše zpráva byla odeslána.',
+    'modal.title' => 'Formulář zpětné vazby',
 
 ];