+++ lang/cs/messages.php
@@ -1,6 +1,0 @@
-<?php
-
-return [
-    'success' => 'Děkujeme za vaši zpětnou vazbu!',
-    'phone_validation' => 'Zadejte prosím platné telefonní číslo s 7-15 číslicemi.',
-];