+++ resources/views/components/link.blade.php
@@ -6,5 +6,5 @@
     {{ $attributes }}
     href="Javascript:void(0)"
     x-data
-    x-on:click.prevent="Livewire.dispatch('openFeedback', {id: '{{ $formId }}'})"
->{{ $slot ?? $formId }}</a>
+    x-on:click.prevent="Livewire.dispatch('open-ig-feedback', {id: '{{ $formId }}'})"
+>{{ $slot->isNotEmpty() ? $slot : __('ig-feedback::layouts.modal.link') }}</a>