+++ resources/views/emails/feedback-plain.blade.php
@@ -6,7 +6,4 @@
 {!! html_entity_decode($field['value']) !!}
 
 @endforeach
-@lang('feedback::layouts.email.sent_from'):
-{{ session('currentPage') ?? '-' }}
-{{ InternetGuru\LaravelCommon\Support\Helpers::getAppInfo() }}
 @endsection